Prediction of score using simple linear regression
The simple linear regression model is designed for predicting the scores of the student one will achieve in the examination if they score for the specific no. of hrs per day. The data consists of no. of scores a student achieved in the past examination and no. of average hours a particular student used to study per day. The model obviuosly depends upon the information provided by the student as it is designed onthe same. The dataset is sample of 25 observation and having two features that is no. of hours a student used to study per day and the percentage is scored in examination. This is an example of simple linear regression with one variable only. As we can see in the codes, certain hypothetical testing is also performed like shapiro test for normality, goldfeld quandt test for homoscedastic testing of the data, Durbin_Watson test and ljungbox test for autocorrelation, etc. along with the graphical analysis of each and every assumption. Finally the model is testing on our own example (here in the code example is what is the percentage a student will score if he study of 9.25 hrs/day).